+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Extend XmlTreeBuilder to allow coordinated build.
+- Introduce BuildCoordinator interface as an extension of DelegateProvider.
+
 ## 0.1.1
 
 - Introduced XmlTreeBuilder with tests and example.